+++ tests/includes/test_only/gdi/test_graphic.cpp
@@ -83,7 +83,7 @@ void TestGraphic::show_mouse_cursor(bool display)
     this->d->gd.show_mouse_cursor(display);
 }
 
-void TestGraphic::set_mouse_cursor_pos(int x, int y)
+void TestGraphic::set_mouse_cursor_pos(uint16_t x, uint16_t y)
 {
     this->d->gd.set_mouse_cursor_pos(x, y);
 }
